Fixed: Now correctly creates the recovery partition when calling the script with only the path of the disk to modify.
The Fix: Removed 'my' before assigning the '$recoverHDBytes' variable inside of the if statement. Otherwise, it would only change the value within the if-then clause and the $recoverHDBytes variable would reset to the previous value (out of scope). This would cause a failure when calling the script and only supplying the path of the disk to add the recovery HD partition to.